NF: Add fullRandom method to MultiStairHandler
`MultiStairHandler` now supports a `fullRandom` randomization method. We
do this by sampling with replacement from the list of all running
staircases on each "pass".

To make this testable, I also added a `randomSeed` kwarg to
`MultiStairHandler`, which is used to seed a `np.random.RandomState`
RNG. All randomization operations are then being done through that
object.

Closes GH-2635.